+++ CHANGELOG.md
@@ -30,6 +30,8 @@ These changes are available on the `master` branch, but have not yet been releas
   ([#2450](https://github.com/Pycord-Development/pycord/pull/2450))
 - Added support for user-installable applications.
   ([#2409](https://github.com/Pycord-Development/pycord/pull/2409)
+- Added support for one-time purchases for Discord monetization.
+  ([#2438](https://github.com/Pycord-Development/pycord/pull/2438))
 
 ### Fixed